+++ plugins/de.cau.cs.kieler.klighd/src/de/cau/cs/kieler/klighd/viewers/ContextViewer.java
@@ -641,7 +641,7 @@ public class ContextViewer implements IViewer, ILayoutRecorder, ISelectionProvid
     /**
      * {@inheritDoc}
      */
-    public void scale(final Object semanticElement, final float scale) {
+    public void scale(final Object semanticElement, final double scale) {
         final EObject diagramElement =
                 getViewContext().getTargetElement(semanticElement, KGraphElement.class);
         if (diagramElement instanceof KNode) {
@@ -652,14 +652,14 @@ public class ContextViewer implements IViewer, ILayoutRecorder, ISelectionProvid
     /**
      * {@inheritDoc}
      */
-    public void scale(final KNode diagramElement, final float scale) {
+    public void scale(final KNode diagramElement, final double scale) {
         currentViewer.scale(diagramElement, scale);
     }
 
     /**
      * {@inheritDoc}
      */
-    public float getScale(final Object semanticElement) {
+    public double getScale(final Object semanticElement) {
         final EObject diagramNode =
                 getViewContext().getTargetElement(semanticElement, KNode.class);
         if (diagramNode instanceof KNode) {
@@ -672,7 +672,7 @@ public class ContextViewer implements IViewer, ILayoutRecorder, ISelectionProvid
     /**
      * {@inheritDoc}
      */
-    public float getScale(final KNode diagramElement) {
+    public double getScale(final KNode diagramElement) {
         return currentViewer.getScale(diagramElement);
     }
